+++ trane/ops/aggregation_ops.py
@@ -43,6 +43,7 @@ class CountAggregationOp(AggregationOpBase):
 
     input_output_types = [("None", "Integer")]
     description = " the number of records"
+    restricted_semantic_tags = {"time_index", "primary_key"}
 
     def label_function(self, dataslice):
         return len(dataslice)
@@ -90,7 +91,7 @@ class MinAggregationOp(AggregationOpBase):
 
 class MajorityAggregationOp(AggregationOpBase):
     input_output_types = [("category", "category")]
-    # input_output_types = [("category", "category"), ("index", "index")]
+    # input_output_types = [("category", "category"), ("primary_key", "primary_key")]
     description = " the majority <{}> in all related records"
 
     def label_function(self, dataslice):
@@ -102,6 +103,7 @@ class MajorityAggregationOp(AggregationOpBase):
 class ExistsAggregationOp(AggregationOpBase):
     input_output_types = [("None", "Boolean")]
     description = " if there exists a record"
+    restricted_semantic_tags = {"time_index", "primary_key"}
 
     def label_function(self, dataslice):
         return len(dataslice) > 0