+++ trane/typing/inference.py
@@ -40,9 +40,17 @@ def _infer_series_schema(series: pd.Series) -> ColumnSchema:
     return ColumnSchema(logical_type=Unknown)
 
 
-def infer_table_meta(df: pd.DataFrame) -> Dict[str, ColumnSchema]:
+def infer_table_meta(
+    df: pd.DataFrame,
+    entity_col=None,
+    time_col=None,
+) -> Dict[str, ColumnSchema]:
     table_meta = {}
     for col in df.columns:
         column_schema = _infer_series_schema(df[col])
         table_meta[col] = column_schema
+    if entity_col:
+        table_meta[entity_col].semantic_tags.add("primary_key")
+    if time_col:
+        table_meta[time_col].semantic_tags.add("time_index")
     return table_meta